Decoding & SciChart: Solved decoding and showing ECG and ICG on the chart
NOTE: This is not tested yet with the real device.

+++ app/src/main/java/nl/hva/vuwearable/models/Measurement.kt
@@ -3,7 +3,7 @@ package nl.hva.vuwearable.models
 data class Measurement(
     val totalBytes: Int,
     val title: String,
-    val formula: ((Double) -> Double)? = null
+    val formula: ((Int) -> Double)? = null
 ) {
-    var value = 0.0
+    var value = 0
 }
